re-doing folder structure for tests
