+++ .vscode/settings.json
@@ -39,8 +39,8 @@
     ],
     "python.sortImports.args": ["-l", "120"],
     "python.testing.pytestArgs": [
-        "${workspaceFolder}/hi-ml/testhiml/",
-        "${workspaceFolder}/hi-ml-azure/testhiml/"
+        "${workspaceFolder}/hi-ml/tests/",
+        "${workspaceFolder}/hi-ml-azure/tests/"
     ],
     "python.testing.unittestEnabled": false,
     "python.testing.nosetestsEnabled": false,
