+++ hi-ml-azure/tests/testazure/test_data/simple/environment.yml.template
