+++ hi-ml-azure/tests/testazure/test_data/simple/hello_world_template.txt
